상품, 타임딜 생성 후 대기열 정책 생성 토큰 발급 후 주문 생성 완료까지 테스트 완료

+++ order-service/src/main/java/com/rushcrew/order_service/infrastructure/adapter/persistence/OrderCommandAdapter.java
@@ -17,7 +17,7 @@ public class OrderCommandAdapter implements OrderCommandPort {
 	private final OrderJpaRepository orderJpaRepository;
 
 	@Override
-	public Integer getTotalPurchasedQuantity(Long userId, UUID productId) {
+	public Long getTotalPurchasedQuantity(Long userId, UUID productId) {
 		return orderJpaRepository.getTotalPurchasedQuantity(userId, productId);
 	}
